[Gold V] Title: 입국심사, Time: 864 ms, Memory: 104968 KB -BaekjoonHub

+++ 백준/Gold/3079. 입국심사/입국심사.java
@@ -17,7 +17,7 @@ public class Main {
 		Arrays.sort(times);
 		
 		long start = 1;
-		long end = (long)times[0] * m;
+		long end = times[0] * m;
 		
 		long mid = start + end / 2;
